Partly Artificial Data on the U.S. Economy

These are partly contrived data from Kmenta (1986), constructed to illustrate estimation of a simultaneous-equation econometric model. The data are an annual time-series for the U.S. economy from 1922 to 1941. The values of the exogenous variables D, and F, and A are real, while those of the endogenous variables Q and P are simulated according to the linear simultaneous equation model fit in the examples.

data("Kmenta", package = "ivreg")

Format

A data frame with 20 rows and 5 columns.

Q

food consumption per capita.

P

ratio of food prices to general consumer prices.

D

disposable income in constant dollars.

F

ratio of preceding year's prices received by farmers to general consumer prices.

A

time in years.

Source

Kmenta, J. (1986) Elements of Econometrics, 2nd ed., Macmillan.

See also

Examples

data("Kmenta", package = "ivreg") 
deq <- ivreg(Q ~ P + D     | D + F + A, data = Kmenta) # demand equation
seq <- ivreg(Q ~ P + F + A | D + F + A, data = Kmenta) # supply equation
summary(deq, tests = TRUE)
#> 
#> Call:
#> ivreg(formula = Q ~ P + D | D + F + A, data = Kmenta)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -3.4305 -1.2432 -0.1895  1.5762  2.4920 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 94.63330    7.92084  11.947 1.08e-09 ***
#> P           -0.24356    0.09648  -2.524   0.0218 *  
#> D            0.31399    0.04694   6.689 3.81e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 1.966 on 17 degrees of freedom
#> Multiple R-Squared: 0.7548,	Adjusted R-squared: 0.726 
#> Wald test: 23.81 on 2 and 17 DF,  p-value: 1.178e-05 
#> 
summary(seq, tests = TRUE)
#> 
#> Call:
#> ivreg(formula = Q ~ P + F + A | D + F + A, data = Kmenta)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -4.8724 -1.2593  0.6415  1.4745  3.4865 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 49.53244   12.01053   4.124 0.000795 ***
#> P            0.24008    0.09993   2.402 0.028785 *  
#> F            0.25561    0.04725   5.410 5.79e-05 ***
#> A            0.25292    0.09966   2.538 0.021929 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 2.458 on 16 degrees of freedom
#> Multiple R-Squared: 0.6396,	Adjusted R-squared: 0.572 
#> Wald test:  10.7 on 3 and 16 DF,  p-value: 0.0004196 
#>