Methods for "ivreg" Objects

Various methods for processing "ivreg" objects; for diagnostic methods, see ivregDiagnostics.

# S3 method for ivreg
coef(object, component = c("stage2", "stage1"), complete = TRUE, ...)

# S3 method for ivreg
vcov(object, component = c("stage2", "stage1"), complete = TRUE, ...)

# S3 method for ivreg
confint(
  object,
  parm,
  level = 0.95,
  component = c("stage2", "stage1"),
  complete = TRUE,
  vcov. = NULL,
  df = NULL,
  ...
)

# S3 method for ivreg
bread(x, ...)

# S3 method for ivreg
estfun(x, ...)

# S3 method for ivreg
vcovHC(x, ...)

# S3 method for ivreg
terms(x, component = c("regressors", "instruments", "full"), ...)

# S3 method for ivreg
model.matrix(
  object,
  component = c("regressors", "projected", "instruments"),
  ...
)

# S3 method for ivreg_projected
model.matrix(object, ...)

# S3 method for ivreg
predict(
  object,
  newdata,
  type = c("response", "terms"),
  na.action = na.pass,
  ...
)

# S3 method for ivreg
print(x, digits = max(3, getOption("digits") - 3), ...)

# S3 method for ivreg
summary(object, vcov. = NULL, df = NULL, diagnostics = NULL, ...)

# S3 method for summary.ivreg
print(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)

# S3 method for ivreg
anova(object, object2, test = "F", vcov. = NULL, ...)

# S3 method for ivreg
update(object, formula., ..., evaluate = TRUE)

# S3 method for ivreg
residuals(
  object,
  type = c("response", "projected", "regressors", "working", "deviance", "pearson",
    "partial", "stage1"),
  ...
)

# S3 method for ivreg
Effect(focal.predictors, mod, ...)

# S3 method for ivreg
formula(x, component = c("complete", "regressors", "instruments"), ...)

# S3 method for ivreg
find_formula(x, ...)

# S3 method for ivreg
Anova(mod, test.statistic = c("F", "Chisq"), ...)

# S3 method for ivreg
linearHypothesis(
  model,
  hypothesis.matrix,
  rhs = NULL,
  test = c("F", "Chisq"),
  ...
)

# S3 method for ivreg
alias(object, ...)

# S3 method for ivreg
qr(x, ...)

# S3 method for ivreg
weights(object, type = c("variance", "robustness"), ...)

Arguments

object, object2, model, mod

An object of class "ivreg".

component

For terms, "regressors", "instruments", or "full"; for model.matrix, "projected", "regressors", or "instruments"; for formula, "regressors", "instruments", or "complete"; for coef and vcov, "stage2" or "stage1".

complete

If TRUE, the default, the returned coefficient vector (for coef()) or coefficient-coevariance matrix (for vcov) includes elements for aliased regressors.

...

arguments to pass down.

parm

parameters for which confidence intervals are to be computed; a vector or numbers or names; the defaiult is all parameters.

level

confidence level; the default is 0.95.

vcov.

Optional coefficient covariance matrix, or a function to compute the covariance matrix, to use in computing the model summary.

df

Optional residual degrees of freedom to use in computing model summary.

x

An object of class "ivreg" or "summary.ivreg".

newdata

Values of predictors for which to obtain predicted values.

type

For predict, one of "response" (the default) or "terms"; for residuals, one of "response" (the default), "projected", "regressors", "working", "deviance", "pearson", or "partial"; type = "working" and "response" are equivalent, as are type = "deviance" and "pearson"; for weights, "variance" (the default) for invariance-variance weights (which is NULL for an unweighted fit) or "robustness" for robustness weights (available for M or MM estimation).

na.action

na method to apply to predictor values for predictions; default is na.pass.

digits

For printing.

diagnostics

Report 2SLS "diagnostic" tests in model summary (default is TRUE). These tests are not to be confused with the regression diagnostics provided elsewhere in the ivreg package: see ivregDiagnostics.

signif.stars

Show "significance stars" in summary output.

test, test.statistic

Test statistics for ANOVA table computed by anova(), Anova(), or linearHypothesis(). Only test = "F" is supported by anova(); this is also the default for Anova() and linearHypothesis(), which also allow test = "Chisq" for asymptotic tests.

formula.

To update model.

evaluate

If TRUE, the default, the updated model is evaluated; if FALSE the updated call is returned.

focal.predictors

Focal predictors for effect plot, see Effect.

hypothesis.matrix, rhs

For formulating a linear hypothesis; see the documentation for linearHypothesis for details.

See also